+++ src/gopls.ts
@@ -1,46 +1,38 @@
 // Routines for working with gopls.
 // const ext = require('ext.js');
 import * as ext from './ext';
+import * as config from './config';
 
-type goplsVersion = {
+type toolVersion = {
     path: string | null;
     version: string | null;
 };
 
-// Obtain the gopls path and version.
-export async function Version(): Promise<goplsVersion> {
-    const ver: goplsVersion = {
-        path: nova.config.get(ext.ns('gopls-path'), 'string'),
+// Obtain the go path and version.
+export async function GoVersion(): Promise<toolVersion> {
+    const ver: toolVersion = {
+        path: ToolPath(config.getString(config.keys.GoPath) ?? 'go'),
         version: null,
     };
-    if (ver.path !== null) {
-        ver.path = ToolPath(ver.path);
-    }
-    if (ver.path !== null) {
-        const out = await ext.exec(ver.path, { args: ['version'] });
-        if (out.status !== 0) {
-            console.error(`gopls return exit code ${out.status}`);
-        } else {
-            const x = out.stdout[0].match(/v\d+\.\d+.\d+/);
-            if (null != x) {
-                ver.version = x[0];
-            }
+    const result = await Go({ args: ['version'] });
+    if (result.status !== 0) {
+        console.error(`go returned an exit code of ${result.status}`);
+    } else {
+        const m = result.stdout[0].match(/go(\d+\.\d+.\d+)/);
+        if (m) {
+            ver.version = m[1];
         }
     }
     return ver;
 }
 
-// Install a version of gopls
-export async function Install(version: string) {
-    console.log(`installing gopls version ${version}`);
-    await Go({
-        args: ['install', `golang.org/x/tools/gopls@${version}`],
-        cwd: '/tmp',
-        env: {
-            GO111MODULE: 'on',
-        },
-    });
-    return Version();
+// Invoke Go
+export function Go(options: ext.ExecOptions): Promise<ext.ExecStatus> {
+    const goPath = ToolPath(config.getString(config.keys.GoPath) ?? 'go');
+    if (null === goPath) {
+        throw 'could not locate go';
+    }
+    return ext.exec(goPath, options);
 }
 
 // Invoke go env VARNAME
@@ -50,6 +42,39 @@ export async function GoEnv(name: string): Promise<string | undefined> {
     });
     const val = result.stdout.join('\n'.trim());
     return val === '' ? undefined : val;
+}
+
+// Obtain the gopls path and version.
+export async function GoplsVersion(): Promise<toolVersion> {
+    const ver: toolVersion = {
+        path: ToolPath(config.getString(config.keys.GoplsPath) ?? 'gopls'),
+        version: null,
+    };
+    if (ver.path !== null) {
+        const out = await ext.exec(ver.path, { args: ['version'] });
+        if (out.status !== 0) {
+            console.error(`gopls return exit code ${out.status}`);
+        } else {
+            const x = out.stdout[0].match(/v(\d+\.\d+.\d+)/);
+            if (null != x) {
+                ver.version = x[1];
+            }
+        }
+    }
+    return ver;
+}
+
+// Install a version of gopls
+export async function GoplsInstall(version: string) {
+    console.log(`installing gopls version ${version}`);
+    await Go({
+        args: ['install', `golang.org/x/tools/gopls@${version}`],
+        cwd: '/tmp',
+        env: {
+            GO111MODULE: 'on',
+        },
+    });
+    return GoplsVersion();
 }
 
 // Return the full path an external tool, or undefined if the
@@ -76,29 +101,4 @@ export function ToolPath(tool: string): string | null {
         return nova.path.join(found, tool);
     }
     return null;
-}
-
-// Obtain the go version.
-export async function GoVersion(): Promise<string | null> {
-    const result = await Go({
-        args: ['version'],
-    });
-    let val = result.stdout.join('\n'.trim());
-    if (val === '') {
-        return null;
-    } else {
-        const m = val.match(/go\d+\.\d+.\d+/);
-        if (!m) {
-            return null;
-        }
-        return m[0];
-    }
-}
-
-export function Go(options: ext.ExecOptions): Promise<ext.ExecStatus> {
-    const goPath = ToolPath('go');
-    if (null === goPath) {
-        throw 'could not locate go';
-    }
-    return ext.exec(goPath, options);
 }