+++ GoNova.novaextension/CHANGELOG.md
@@ -1,7 +1,12 @@
+## Version 1.2
+
+* Added an extension preference for the path to the `go` command, allowing the use different versions.
+* Added workspace preferences for the `go` and `gopls` command paths, and well as format-on-save and organize imports on save.
+
 ## Version 1.1
 
-Switch to tree-sitter syntax. Requires Nova 10.
+* Switch to tree-sitter syntax. Requires Nova 10.
 
 ## Version 1.0
 
-Initial release
+* Initial release